feat: add EditorProvider and hooks to provide granular updates
