+++ packages/react/src/EditorProvider.ts
@@ -1,0 +1,11 @@
+import { Editor } from '@tiptap/react'
+import { createContext } from 'react'
+
+export const EditorContext = createContext<Editor | null>(null)
+
+const { Provider, Consumer } = EditorContext
+
+export {
+  Provider as EditorProvider,
+  Consumer as EditorConsumer,
+}
