+++ packages/react/src/index.ts
@@ -3,6 +3,8 @@ export * from './BubbleMenu'
 export { Editor } from './Editor'
 export * from './FloatingMenu'
 export * from './useEditor'
+export * from './useCurrentEditor'
+export * from './EditorProvider'
 export * from './ReactRenderer'
 export * from './ReactNodeViewRenderer'
 export * from './EditorContent'
